restructure secureprotol dir
