+++ federatedml/ensemble/boosting/boosting_core/hetero_boosting.py
@@ -1,3 +1,22 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+#
+#  Copyright 2019 The FATE Authors. All Rights Reserved.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+#
+
 from abc import ABC
 import abc
 from federatedml.ensemble.boosting.boosting_core import Boosting
@@ -46,7 +65,8 @@ class HeteroBoosting(Boosting, ABC):
             self.encrypter.generate_key(self.encrypt_param.key_length)
         elif self.encrypt_param.method.lower() == consts.ITERATIVEAFFINE.lower():
             self.encrypter = IterativeAffineEncrypt()
-            self.encrypter.generate_key(self.encrypt_param.key_length)
+            self.encrypter.generate_key(key_size=self.encrypt_param.key_length,
+                                        randomized=self.encrypt_param.randomized)
         else:
             raise NotImplementedError("encrypt method not supported yes!!!")
