+++ federatedml/framework/hetero/procedure/two_parties_paillier_cipher.py
@@ -16,7 +16,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from federatedml.secureprotol.encrypt import PaillierEncrypt
+from federatedml.secureprotol.he.encrypt import PaillierEncrypt
 from federatedml.util import consts
