+++ federatedml/framework/homo/blocks/random_padding_cipher.py
@@ -17,7 +17,7 @@ from federatedml.framework.homo.blocks import uuid_generator, diffie_hellman
 from federatedml.framework.homo.blocks.base import HomoTransferBase
 from federatedml.framework.homo.blocks.diffie_hellman import DHTransVar
 from federatedml.framework.homo.blocks.uuid_generator import UUIDTransVar
-from federatedml.secureprotol.encrypt import PadsCipher
+from federatedml.secureprotol.he.encrypt import PadsCipher
 from federatedml.util import LOGGER
 from federatedml.util import consts
