+++ federatedml/secureprotol/he/encrypt.py
@@ -21,12 +21,12 @@ from Cryptodome import Random
 from Cryptodome.PublicKey import RSA
 
 from federatedml.secureprotol import gmpy_math
-from federatedml.secureprotol.affine import AffineCipher
-from federatedml.secureprotol.fate_paillier import PaillierKeypair
+from federatedml.secureprotol.he.affine import AffineCipher
+from federatedml.secureprotol.he.fate_paillier import PaillierKeypair
 from federatedml.secureprotol.random import RandomPads
 
 
-from federatedml.secureprotol.iterative_affine import IterativeAffineCipher
+from federatedml.secureprotol.he.iterative_affine import IterativeAffineCipher
 
 
 class Encrypt(object):
@@ -254,8 +254,10 @@ class IterativeAffineEncrypt(SymmetricEncrypt):
     def __init__(self):
         super(IterativeAffineEncrypt, self).__init__()
 
-    def generate_key(self, key_size=1024, key_round=5):
-        self.key = IterativeAffineCipher.generate_keypair(key_size=key_size, key_round=key_round)
+    def generate_key(self, key_size=1024, key_round=5, randomized=False):
+        self.key = IterativeAffineCipher.generate_keypair(key_size=key_size,
+                                                          key_round=key_round,
+                                                          randomized=randomized)
 
     def encrypt(self, plaintext):
         if self.key is not None:
