+++ federatedml/secureprotol/he/fate_paillier.py
@@ -16,8 +16,7 @@
 #  limitations under the License.
 #
 
-from collections.abc import Mapping
-from federatedml.secureprotol.fixedpoint import FixedPointNumber
+from federatedml.secureprotol.he.fixedpoint import FixedPointNumber
 from federatedml.secureprotol import gmpy_math
 import random
